Draft implementation for custom Cargo profiles
This allows creating custom profiles that inherit from other profiles.

For example, one can have a release-lto profile that looks like this:

    [profile.custom.release-lto]
    inherits = "release"
    lto = true

The profile name will also carry itself into the output directory name
so that the different build outputs can be cached independently from
one another.

So in effect, at the `target` directory, a name will be created for
the new profile, in addition to the 'debug' and 'release' builds:

```
    $ cargo build --profile release-lto
    $ ls -l target
    debug release release-lto
```

Check list (or, what is missing):

- [ ] Implement tests
- [ ] Write documentation
- [ ] Replace panic!() calls with better error handling?
- [ ] See whether `--profile` can be passed to commands other than `build`
- [ ] See whether it makes sense to override profiles other than `dev`
and `release`
- [ ] Probably more refactoring is needed to insert this cleanly
- [ ] Cleanups